requirements.txt
alembic==0.9.9
aniso8601==3.0.0
asn1crypto==0.24.0
azure-common==1.1.10
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.2
azure-nspkg==2.0.0
azure-storage-common==1.1.0
azure-storage-nspkg==3.0.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2.2
Flask==1.0
Flask-Migrate==2.1.1
flask-restplus==0.10.1
Flask-SQLAlchemy==2.3.2
futures==3.1.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
protobuf==3.5.2.post1
pycparser==2.18
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2018.4
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.7
urllib3==1.22
Werkzeug==0.14.1
